Big Changes **PLZ READ**
User can log in, move to units tab, select a unit and add it to there
list. They can then lost into that unit. Unit is currently not saved to
that user however so must be added each time.  Still a bit buggy though
as you have to hit send twice to post and it wont post exactly right.
You can no longer just hit "LOGIN" to start the app. Account must first
be created and only the user name has to be used, pw not implemented
yet.